Possibly final structure for the language server folders
