+++ Code/Language Server/m/src/m/formatting/YAMLFormat.java
@@ -2,7 +2,7 @@ package m.formatting;
 
 import m.yaml.Stream;
 
-public class YAMLFormat extends GenericFormatter
+public class YAMLFormat extends FormattingHelper
 {
 	@Override
 	protected void format(Object o)
