--- Code/Language Server/m/src/m/library/Type.java
@@ -1,6 +1,0 @@
-package m.library;
-
-public interface Type 
-{
-
-}
